requirements.en.txt
bcrypt==3.1.4
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
decorator==4.1.2
Django==1.10.6
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.1
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.5.3
pymodm==0.4.0
pymongo==3.5.1
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
